style(provd): split mocks into multiple files
